feat: Add a confirmation to delete a journal entry #4308

+++ resources/js/components/journal/partials/JournalContentEntry.vue
@@ -85,10 +85,12 @@ export default {
     },
 
     trash() {
-      axios.delete('journal/' + this.entry.id)
-        .then(response => {
-          this.$emit('deleteJournalEntry', this.journalEntry.id);
-        });
+      if( confirm(this.$t('journal.delete_confirmation')) ) {
+        axios.delete('journal/' + this.entry.id)
+          .then(response => {
+            this.$emit('deleteJournalEntry', this.journalEntry.id);
+          });
+      }
     },
 
     compiledMarkdown (text) {
